+++ datafaker/proposers/base.py
@@ -311,18 +311,11 @@ class Buckets:
         join_tables: list[Table] | None = None,
     ):
         """Initialise a Buckets object."""
-        bottom = mean - 2 * stddev
-        width = stddev / 2
-        query = select(
-            func.count(column).label("f"),  # pylint: disable=not-callable
-            ((func.floor(column) - bottom) / width).label("b"),
-        ).select_from(table)
-        for jt in join_tables or []:
-            query = query.join(jt)
+        self.mean = mean
+        self.stddev = stddev
+        query = self._get_bucket_query(table, column, join_tables)
         with engine.connect() as connection:
-            raw_buckets = connection.execute(
-                query.group_by("b")
-            )
+            raw_buckets = connection.execute(query.group_by("b"))
             self.buckets: Sequence[int] = [0] * 10
             for rb in raw_buckets:
                 try:
@@ -337,8 +330,23 @@ class Buckets:
                     # catches errors if SQLAlchemy returns something that
                     # isn't a number for some other unknown reason.
                     pass
-        self.mean = mean
-        self.stddev = stddev
+
+    def _get_bucket_query(
+        self,
+        table: Table | Join,
+        column: Any,
+        join_tables: list[Table] | None,
+    ) -> Any:
+        """Make the query for half-sd buckets."""
+        bottom = self.mean - 2 * self.stddev
+        width = self.stddev / 2
+        query = select(
+            func.count(column).label("f"),  # pylint: disable=not-callable
+            ((func.floor(column) - bottom) / width).label("b"),
+        ).select_from(table)
+        for jt in join_tables or []:
+            query = query.join(jt)
+        return query
 
     @classmethod
     def make_buckets(
